azure-function-dbt/requirements.txt
adlfs==2025.8.0
agate==1.9.1
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
asttokens==3.0.1
attrs==25.4.0
azure-core==1.38.0
azure-datalake-store==0.0.53
azure-identity==1.25.1
azure-storage-blob==12.28.0
azure-storage-file-datalake==12.23.0
babel==2.18.0
beautifulsoup4==4.14.3
bs4==0.0.2
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
comm==0.2.3
cryptography==46.0.3
curl-cffi==0.14.0
daff==1.4.2
dbt-adapters==1.22.5
dbt-common==1.37.2
dbt-core==1.11.2
dbt-duckdb==1.10.0
dbt-extractor==0.6.0
dbt-protos==1.0.427
dbt-semantic-interfaces==0.9.0
debugpy==1.8.19
decorator==5.2.1
deepdiff==7.0.1
duckdb==1.4.4
executing==2.2.1
frozenlist==1.8.0
fsspec==2026.1.0
google-api-core==2.29.0
google-auth==2.47.0
google-cloud-core==2.5.0
google-cloud-storage==3.8.0
google-crc32c==1.8.0
google-resumable-media==2.8.0
googleapis-common-protos==1.72.0
idna==3.11
importlib-metadata==8.7.1
ipykernel==7.1.0
ipython==9.9.0
ipython-pygments-lexers==1.1.1
isodate==0.6.1
jedi==0.19.2
jinja2==3.1.6
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter-client==8.8.0
jupyter-core==5.9.1
leather==0.4.1
markupsafe==3.0.3
mashumaro==3.14
matplotlib-inline==0.2.1
more-itertools==10.8.0
msal==1.34.0
msal-extensions==1.3.1
msgpack==1.1.2
multidict==6.7.1
nest-asyncio==1.6.0
networkx==3.6.1
numpy==2.4.1
ordered-set==4.1.0
packaging==25.0
pandas==3.0.0
parsedatetime==2.6
parso==0.8.5
pathspec==0.12.1
pexpect==4.9.0
platformdirs==4.5.1
prompt-toolkit==3.0.52
propcache==0.4.1
proto-plus==1.27.0
protobuf==6.33.5
psutil==7.2.1
ptyprocess==0.7.0
pure-eval==0.2.3
pyarrow==23.0.0
pyasn1==0.6.1
pyasn1-modules==0.4.2
pycparser==2.23
pydantic==1.10.26
pygments==2.19.2
pyjwt==2.10.1
python-dateutil==2.9.0.post0
python-slugify==8.0.4
pytimeparse==1.1.8
pytz==2025.2
pyyaml==6.0.3
pyzmq==27.1.0
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
rsa==4.9.1
six==1.17.0
snowplow-tracker==1.1.0
soupsieve==2.8.1
sqlparse==0.5.4
stack-data==0.6.3
text-unidecode==1.3
tornado==6.5.4
traitlets==5.14.3
typing-extensions==4.15.0
urllib3==2.6.3
wcwidth==0.2.14
yarl==1.22.0
zipp==3.23.0
